requirements.txt
psutil
opencv-python
Pillow 
pygame
ffpyplayer
py-cpuinfo